improved plotting and a bug fix
There are now a gazillion options to MOHGP.plot. Use at your own risk

in col_mix, some of the code for splitting a cluster could occasionally
results in an error, if the cluster didn;t have sufficient samples in to
divide into two. this should be fixed now (the cluster doesn't get
split)